Module Abstractions.Network

type links = int array

The type representing the neighbors of a node.

type network_links = links array

The type representing the neighbors of each node in the simulation.

type region = int

The type representing the region of a node.

type regions = int array

The type representing the region assigned to each node in the simulation

module type Network = sig ... end

The type of the Network abstraction. Provides primitives for creating the network topology and for abstracting message exchanges.

module Make (Events : Simulator.Events.Event) (Queue : Simulator.Events.EventQueue with type ev = Events.t) (Message : Simulator.Events.Message with type t = Events.msg) : Network with type msg = Events.msg

Functor that creates an implementation for the Network, given the Events, EventQueue and Message modules.