Module Network.Make

Functor that creates an implementation for the Network, given the Events, EventQueue and Message modules.

Parameters

module Events : Simulator.Events.Event

module Queue : Simulator.Events.EventQueue with type ev = Events.t

module Message : Simulator.Events.Message with type t = Events.msg

Signature

type msg = Events.msg

The type of the messages being exchanged

module MessagesExchanged : Simulator.Statistics.Stats

Stats module to keep track of the number of messages exchanged during the simulation

module MegabytesExchanged : Simulator.Statistics.Stats

Stats module to keep track of the total megabytes exchanged during the simulation

val send : int -> int -> msg -> unit

Primitive for point-to-point message exchange.

parameter sender_node_id
the id of the node sending the message

parameter target_node_id
the id of the node receiving the message

parameter message
the contents of the message being sent

val send_to_neighbors : int -> msg -> unit

Primitive to send a message to all neighbors of a given node

parameter sender_node_id
the id of the node sending the message

parameter message
the contents of the message being sent

val gossip : int -> msg -> unit

Primitive to abstract the gossip protocol. Gossips a message to every node in the network. Used for slightly more relaxed scenarios, since it assumes that limited-bandwidth=false.

parameter sender_node_id
the id of the node sending the message

parameter message
the contents of the message being sent

val get_neighbours : int -> links

Primitive to obtain the neighbors of a given node.

parameter node_id
the id of the node whose neighbors we want

val clear : unit -> unit

Primitive to reset all data stored by the network.

val get_links : unit -> network_links

Primitive to get the links assigned to each node in the network.

val get_regions : unit -> regions

Primitive to get the region assigned to each node in the network.